Revert "Introduce base entity for ping (#104197)"
This reverts commit 6f54aaf564d9cbca439fcac301bbb57bd7c5ecad.
